+++ src/tee/components/MarkdownEditor.tsx
@@ -318,6 +318,10 @@ export function MarkdownEditor({
           if (evt.key === "\\" && (evt.metaKey || evt.ctrlKey)) {
             return;
           }
+          // Let cmd-g thru for grouping annotations
+          if (evt.key === "g" && (evt.metaKey || evt.ctrlKey)) {
+            return;
+          }
           evt.stopPropagation();
         }}
       />
